+++ pymars/tests/test_soln2cti.py
@@ -188,3 +188,112 @@ class TestWrite:
             new_solution = ct.Solution(output)
 
         assert compare_models(solution, new_solution)
+
+    def test_multiple_pressure_dependent_reactions(self):
+        """Test appropriate writing of multiple pressure dependent reactions.
+
+        We can encounter multiple instances of the same reaction, but with
+        different explicit third-body species.
+        """
+        species_h = ct.Species.fromCti(
+        '''species(name='h', atoms='H:1',
+        thermo=(NASA([200.00, 1000.00],
+                     [ 2.50000000E+00,  0.00000000E+00,  0.00000000E+00,
+                       0.00000000E+00,  0.00000000E+00,  2.54736600E+04,
+                      -4.46682850E-01]),
+                NASA([1000.00, 6000.00],
+                     [ 2.50000000E+00,  0.00000000E+00,  0.00000000E+00,
+                       0.00000000E+00,  0.00000000E+00,  2.54736600E+04,
+                      -4.46682850E-01])))'''
+        )
+        species_o2 = ct.Species.fromCti(
+        '''species(name='o2', atoms='O:2',
+        thermo=(NASA([200.00, 1000.00],
+                     [ 3.78245636E+00, -2.99673416E-03,  9.84730201E-06,
+                      -9.68129509E-09,  3.24372837E-12, -1.06394356E+03,
+                       3.65767573E+00]),
+                NASA([1000.00, 6000.00],
+                     [ 3.66096065E+00,  6.56365811E-04, -1.41149627E-07,
+                       2.05797935E-11, -1.29913436E-15, -1.21597718E+03,
+                       3.41536279E+00])))'''
+        )
+        species_ho2 = ct.Species.fromCti(
+            '''species(name='ho2', atoms='H:1 O:2',
+        thermo=(NASA([200.00, 1000.00],
+                     [ 4.30179807E+00, -4.74912097E-03,  2.11582905E-05,
+                      -2.42763914E-08,  9.29225225E-12,  2.64018485E+02,
+                       3.71666220E+00]),
+                NASA([1000.00, 5000.00],
+                     [ 4.17228741E+00,  1.88117627E-03, -3.46277286E-07,
+                       1.94657549E-11,  1.76256905E-16,  3.10206839E+01,
+                       2.95767672E+00])))'''
+        )
+        species_ar = ct.Species.fromCti(
+            '''species(name='ar', atoms='Ar:1',
+        thermo=(NASA([200.00, 1000.00],
+                     [ 2.50000000E+00,  0.00000000E+00,  0.00000000E+00,
+                       0.00000000E+00,  0.00000000E+00, -7.45375000E+02,
+                       4.37967491E+00]),
+                NASA([1000.00, 6000.00],
+                     [ 2.50000000E+00,  0.00000000E+00,  0.00000000E+00,
+                       0.00000000E+00,  0.00000000E+00, -7.45375000E+02,
+                       4.37967491E+00])))'''
+        )
+        species_he = ct.Species.fromCti(
+            '''species(name='he', atoms='He:1',
+        thermo=(NASA([200.00, 1000.00],
+                     [ 2.50000000E+00,  0.00000000E+00,  0.00000000E+00,
+                       0.00000000E+00,  0.00000000E+00, -7.45375000E+02,
+                       9.28723974E-01]),
+                NASA([1000.00, 6000.00],
+                     [ 2.50000000E+00,  0.00000000E+00,  0.00000000E+00,
+                       0.00000000E+00,  0.00000000E+00, -7.45375000E+02,
+                       9.28723974E-01])))'''
+        )
+        species_h2o = ct.Species.fromCti(
+            '''species(name='h2o', atoms='H:2 O:1',
+        thermo=(NASA([200.00, 1000.00],
+                     [ 4.19863520E+00, -2.03640170E-03,  6.52034160E-06,
+                      -5.48792690E-09,  1.77196800E-12, -3.02937260E+04,
+                      -8.49009010E-01]),
+                NASA([1000.00, 6000.00],
+                     [ 2.67703890E+00,  2.97318160E-03, -7.73768890E-07,
+                       9.44335140E-11, -4.26899910E-15, -2.98858940E+04,
+                       6.88255000E+00])))'''
+        )
+        full_species = [species_h, species_o2, species_ho2, species_ar, species_he, species_h2o]
+        
+        R1 = ct.FalloffReaction.fromCti(
+        "units(length='cm', quantity='mol')\n" +
+        '''falloff_reaction('h + o2 (+ M) <=> ho2 (+ M)',
+            kf=[4.650000e+12, 0.44, 0.0],
+            kf0=[1.737000e+19, -1.23, 0.0],
+            efficiencies='he:0.0 h2o:10.0 ar:0.0',
+            falloff=Troe(A=0.67, T3=1e-30, T1=1e+30, T2=1e+30))'''
+        )
+        R2 = ct.FalloffReaction.fromCti(
+        "units(length='cm', quantity='mol')\n" +
+        '''falloff_reaction('h + o2 (+ ar) <=> ho2 (+ ar)',
+            kf=[4.650000e+12, 0.44, 0.0],
+            kf0=[6.810000e+18, -1.2, 0.0],
+            falloff=Troe(A=0.7, T3=1e-30, T1=1e+30, T2=1e+30))'''
+        )
+        R3 = ct.FalloffReaction.fromCti(
+        "units(length='cm', quantity='mol')\n" +
+        '''falloff_reaction('h + o2 (+ he) <=> ho2 (+ he)',
+            kf=[4.650000e+12, 0.44, 0.0],
+            kf0=[9.192000e+18, -1.2, 0.0],
+            falloff=Troe(A=0.59, T3=1e-30, T1=1e+30, T2=1e+30))'''
+        )
+
+        solution = ct.Solution(
+            species=full_species, 
+            reactions=[R1, R2, R3],
+            thermo='IdealGas', kinetics='GasKinetics'
+            )
+        
+        with TemporaryDirectory() as temp_dir:
+            output = write(solution, 'test.cti', path=temp_dir)
+            new_solution = ct.Solution(output)
+
+        assert compare_models(solution, new_solution)